+++ TodoBack/Repos/Interfaces/ITodoItemRepository.cs
@@ -11,10 +11,10 @@ namespace TodoBack.Repos.Interfaces
     public interface ITodoItemRepository
     {
         Task<TodoItem> GetByIdAsync(long id);
-        Task<IEnumerable<TodoItem>> GetByUserIdAsync(int userId);
+        Task<IEnumerable<TodoItem>> GetByUserIdAsync(long userId);
         Task  AddAsync(TodoItem item);
         Task UpdateAsync(TodoItem item);
-        Task<bool> DeleteAsync(int id);
-        Task<int> DeleteByUserIdAsync(int userId);
+        Task<bool> DeleteAsync(long id);
+        Task<int> DeleteByUserIdAsync(long userId);
     }
 }
